build: change references to S3 to memory and set myself as author
The upstream fork of this project was a an S3-based VFS, and so all references
to S3 should be changed to memory.

Also adding myself as the (first) author. The DIT is only I guess "historical"
author since they wrote the upstream fork. But also, the DIT doesn't exist any
more - not quite sure how to handle that have to admit.