fix: more robust locking
The main issues from the previous code was a misunderstanding of how locking
should work. The lock level is per file object and _not_ to be shared between
file objects on the same file in the VFS. But the total numbers of readers and
writers _is_ shared.

I wonder if this is a bit too strict - once a reserved lock is granted it will
block other shared locks, even though https://www.sqlite.org/lockingv3.html
doesn't specify that. However, I think it's safer in the short term at least to
do this.